% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test_cpp.R
\name{function_test_cpp}
\alias{function_test_cpp}
\title{Step for permuting two elements in a given sequence.}
\usage{
function_test_cpp(x, ...)
}
\arguments{
\item{x}{vector}
}
\value{
sample of elements of the vector
}
\description{
Step for permuting two elements in a given sequence.
}
