% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_preprocessing.R
\name{GatherPreprocessingDF}
\alias{GatherPreprocessingDF}
\title{Gather preprocessing function}
\usage{
GatherPreprocessingDF(formula, envir = new.env(), submodel = "choice")
}
\arguments{
\item{formula}{string formula with the effects for the model}

\item{envir}{environment were dependentEvents, nodes and net0 are located.}
}
\value{
listExpandedDF (see \code{\link[goldfish:estimate]{goldfish::estimate()}} for more)
}
\description{
TO DO: include Rate model / REM specification
}
