% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{getAuxDfE}
\alias{getAuxDfE}
\title{Get auxiliary data frame for tie \eqn{e}}
\usage{
getAuxDfE(auxDf, sender, receiver)
}
\arguments{
\item{auxDf}{data frame from getKelMeMatrix.}

\item{sender}{sender node}

\item{receiver}{receiver node}
}
\value{
list of auxDfE (data frame).
}
\description{
Get auxiliary data frame for tie \eqn{e}
}
