% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{getKelMeMatrix}
\alias{getKelMeMatrix}
\title{Get auxiliary matrix \eqn{K_{el}} and \eqn{m_e} for MCMC steps.}
\usage{
getKelMeMatrix(seq, actDfnodesLab)
}
\arguments{
\item{seq}{data frame with the sequence of events up to time \eqn{t}.}

\item{actDfnodes}{vector with the labels of the actors of the network.}
}
\value{
list with matrix \eqn{K_{el}}, \eqn{K_{e,l>1}}, \eqn{m_e} and
auxiliary data frame
}
\description{
Get auxiliary matrix \eqn{K_{el}} and \eqn{m_e} for MCMC steps.
}
\examples{
seq <- data.frame(
  "sender" = c(1, 2, 3), "receiver" = c(3, 2, 1),
  "replace" = c(0, 1, 1)
)
getKelMeMatrix(seq, actDfnodes = c(1, 2, 3))
}
