% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{getlogLikelihood}
\alias{getlogLikelihood}
\title{log-likelihood computation: Multi-Core with time}
\usage{
getlogLikelihood(
  seq,
  actDfnodes,
  net0,
  fixedparameters,
  parameters,
  initTime,
  endTime,
  formula
)
}
\value{
vector of loglikelihoods
}
\description{
given parameters, sequences, formula, initial network and
multi-core elements, computes log-likelihood of all the sequences.
}
