% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{getlogLikelihoodMC}
\alias{getlogLikelihoodMC}
\title{log-likelihood computation: Multi-Core with time}
\usage{
getlogLikelihoodMC(
  indexCore,
  seqsEM,
  beta,
  fixedparameters,
  splitIndicesPerCore,
  initTime,
  endTime,
  actDfnodes = actDfnodes,
  net0 = net0,
  formula = formula,
  temp
)
}
\value{
vector of loglikelihoods
}
\description{
given parameters, sequences, formula, initial network and
multi-core elements, computes log-likelihood of all the sequences.
}
