% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{getlogLikelihoodRate}
\alias{getlogLikelihoodRate}
\title{log-likelihood computation: Multi-Core with time}
\usage{
getlogLikelihoodRate(seq, actDfnodes, parameters, initTime, endTime, temp)
}
\value{
vector of loglikelihoods
}
\description{
given parameters, sequences, formula, initial network and
multi-core elements, computes log-likelihood of all the sequences.
}
