% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{getpDoAugment}
\alias{getpDoAugment}
\title{Get probability of doing a augmenting step.}
\usage{
getpDoAugment(
  gammaEplus,
  gammaPlus,
  p,
  m,
  me,
  sender,
  receiver,
  typeA,
  indexSR,
  pAug
)
}
\description{
Get probability of doing a augmenting step.
}
