% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{getpDoPerm}
\alias{getpDoPerm}
\title{Get probability of doing a permutation step.}
\usage{
getpDoPerm(
  probVec,
  probVec2,
  e1,
  e2,
  sender1,
  receiver1,
  sender2,
  receiver2,
  seq
)
}
\description{
Get probability of doing a permutation step.
}
