% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{getpDoShort}
\alias{getpDoShort}
\title{Get probability of doing a shortening step.}
\usage{
getpDoShort(
  gammaEminus,
  gammaMinus,
  p,
  uniqueR,
  sender,
  receiver,
  typeS,
  pShort
)
}
\description{
Get probability of doing a shortening step.
}
