% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{logLikelihood}
\alias{logLikelihood}
\title{log-likelihood computation}
\usage{
logLikelihood(
  listExpandedDFChoice,
  listExpandedDFRate,
  beta,
  theta,
  initTime,
  endTime
)
}
\arguments{
\item{beta}{list of choice parameters with elements Crea and Del.}

\item{listExpandedDF}{list returned from function GatherPreprocessingDF}
}
\value{
loglikelihood (creation + deletion of events models)
}
\description{
log-likelihood computation
}
