% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{logLikelihoodMCTemp}
\alias{logLikelihoodMCTemp}
\title{log-likelihood computation: Multi-Core with temperatures (parallel tempering)}
\usage{
logLikelihoodMCTemp(
  indexCore,
  permut,
  beta,
  theta,
  initTime,
  endTime,
  splitIndicesPerCore,
  actDfnodes = actDfnodes,
  net0 = net0,
  formula = formula,
  temp
)
}
\value{
vector of loglikelihoods
}
\description{
given parameters, sequences, formula, initial network and
multi-core elements, computes log-likelihood of all the sequences.
}
