% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{logLikRate}
\alias{logLikRate}
\title{log-likelihood computation with time information}
\usage{
logLikRate(expandedDF, theta)
}
\arguments{
\item{theta}{list of rate parameters with elements Crea and Del}

\item{listExpandedDF}{list returned from function GatherPreprocessingDF}
}
\value{
loglikelihood (creation + deletion of events models)
}
\description{
log-likelihood computation with time information
}
