% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_likelihood.R
\name{logLikTime}
\alias{logLikTime}
\title{log-likelihood computation with time information}
\usage{
logLikTime(logLikRate, initTime, endTime)
}
\arguments{
\item{logLikRate}{list returned from logLik}

\item{initTime}{initial time of observation}

\item{endTime}{end time of observation}
}
\value{
loglikelihood (creation + deletion of events models)
}
\description{
log-likelihood computation with time information
}
