% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{MCMC}
\alias{MCMC}
\title{MCMC function (simple version with augmentation/shortening/permutation)}
\usage{
MCMC(
  nmax,
  seq,
  H,
  actDfnodes,
  formula,
  net0,
  beta,
  theta,
  initTime,
  endTime,
  burnIn = TRUE,
  maxIter = 10000,
  seqIter = 50,
  pShort = 0.35,
  pAug = 0.35,
  pPerm = 0.3
)
}
\arguments{
\item{nmax}{integer, number of sequences to be returned.}

\item{seq}{data frame, sequence from which start/continue MCMC chain.}

\item{H}{integer, hamming distance between net0 and net1.}

\item{actDfnodes}{object of type ´nodes´ from goldfish.}

\item{formula}{formula, the one used in goldfish model.}

\item{net0}{matrix, initial observed network.}

\item{beta}{list, estimator of parameters of the model (creation and deletion).}

\item{burnIn}{boolean, indicates if burn in must be performed.}

\item{maxIter}{integer, maximum number of steps of the MCMC chain.}

\item{seqIter}{integer, number of steps between selected sequences.}

\item{pShort}{float, probability of shortening step.}

\item{pAug}{float, probability of augmenting step.}

\item{pPerm}{float, probability of permutation step.}
}
\value{
permut, list of sequences.
}
\description{
Computes an MCMC chain and returns nmax sequences from the MCMC chain.
}
