% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{MCMC_rate}
\alias{MCMC_rate}
\title{MCMC_rate (simple version with augmentation/shortening/permutation)}
\usage{
MCMC_rate(
  nmax,
  seqInit,
  H,
  actDfnodes,
  formula,
  net0,
  beta,
  theta,
  fixedparameters,
  initTime,
  endTime,
  burnIn = TRUE,
  burnInIter = 500,
  maxIter = 10000,
  thin = 50,
  pAug = 0.35,
  pShort = 0.35,
  pPerm = 0.3,
  k = 5,
  index = 1,
  out_file_names = c("out_PT_acceptDF", "out_PT_mcmcDiagDF"),
  changePPerm = FALSE
)
}
\arguments{
\item{nmax}{integer, number of sequences to be returned.}

\item{seqInit}{data frame, sequence from which start/continue MCMC chain.}

\item{H}{integer, Hamming distance between net0 and net1.}

\item{actDfnodes}{object of type ´nodes´ from goldfish.}

\item{formula}{formula, the one used in goldfish model.}

\item{net0}{matrix, initial observed network.}

\item{beta}{data frame, initial values of parameters of the choice model (creation and deletion).}

\item{theta}{data frame, initial values of parameters of the rate model (creation and deletion).}

\item{fixedparameters}{data frame, fixed values for parameters of the
choice model (creation and deletion), set as NA if parameter is not fixed.}

\item{initTime}{time of initial observation.}

\item{endTime}{time of final observation.}

\item{burnIn}{boolean, indicates if burn-in must be performed.}

\item{burnInIter}{integer, number of burn-in iterations.}

\item{maxIter}{integer, maximum number of steps of the MCMC chain.}

\item{thin}{integer, number of steps between selected sequences.}

\item{pAug}{float, probability of augmenting step.}

\item{pShort}{float, probability of shortening step.}

\item{pPerm}{float, probability of permutation step.}

\item{k}{integer, number or permutations after an augmentation, shortening or permutaion mutation.}

\item{index}{integer, auxiliar value for writing tables in txt files.
Set different that 1 if files already exist.}

\item{out_file_names}{names for saving tables with acceptance of mutations,
and loglikelihoods and score vectors. Do NOT add .txt extension.}

\item{changepPerm}{boolean, allows for adaptative permutation mutation probability.}
}
\value{
list of sampled sequences, seqsEM, and final state of the chain, resstepPT.
}
\description{
Computes an MCMC chain and returns nmax sequences from the MCMC chain.
}
