% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_parameters.R
\name{newtonraphsonStepRate}
\alias{newtonraphsonStepRate}
\title{newtonraphsonStepRate}
\usage{
newtonraphsonStepRate(parameters, seqsEM)
}
\arguments{
\item{x}{values of the parameters}

\item{se}{standard errors of the parameters}

\item{w}{values of the weights}
}
\value{
value of the averaged mean and the standard error
}
\description{
newtonraphsonStepRate
}
