% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_parameters.R
\name{parameters}
\alias{parameters}
\title{Function that computes goldfish estimation of parameters for a given sequence}
\usage{
parameters(seq, actDfnodes. = actDfnodes, net0. = net0, formula. = formula)
}
\arguments{
\item{seq}{sequence of values}

\item{actDfnodes.}{nodes}

\item{net0.}{initial network}

\item{formula.}{formula of the model}
}
\value{
list with goldfish estimators and their standard errors
}
\description{
Function that computes goldfish estimation of parameters for a given sequence
}
