% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_parameters.R
\name{parametersMC}
\alias{parametersMC}
\title{Estimation of parameters: Multi-Core}
\usage{
parametersMC(
  indexCore,
  splitIndicesPerCore,
  permut = permut,
  actDfnodes. = actDfnodes,
  net0. = net0,
  formula. = formula
)
}
\value{
list with goldfish estimators and their standard errors for each permuation
}
\description{
function that paralelizes computation of parameters on all the sequences using goldfish.
}
