% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{PT_MCMC}
\alias{PT_MCMC}
\title{Parallel tempering MCMC function (augmentation/shortening)}
\usage{
PT_MCMC(
  nmax,
  nPT,
  seqsPT,
  H,
  actDfnodes,
  formula,
  net0,
  beta,
  theta,
  fixedparameters,
  initTime,
  endTime,
  burnIn = TRUE,
  burnInIter = 500,
  maxIter = 10000,
  thin = 50,
  T0 = 100,
  nStepExch = 10,
  pAug = 0.35,
  pShort = 0.35,
  pPerm = 0.3,
  k = 5,
  cl = cl,
  num_cores = num_cores
)
}
\arguments{
\item{nmax}{integer, number of sequences to sampled.}

\item{nPT}{integer, number of chains in the parallel tempering algorithm.}

\item{H}{integer, hamming distance between net0 and net1.}

\item{actDfnodes}{object of type ´nodes´ from goldfish.}

\item{formula}{formula, the one used in goldfish model.}

\item{net0}{matrix, initial observed network.}

\item{beta}{list, estimator of parameters of the model (creation and deletion).}

\item{burnIn}{boolean, indicates if burn in must be performed.}

\item{maxIter}{integer, maximum number of steps of the MCMC chain.}

\item{thin}{integer, number of steps between selected sequences.}

\item{T0}{initial maximal tempetature for the parallale tempering scheme/
simulated annealing}

\item{nStepExch}{number of mutation steps before a exchange}

\item{pAug}{float, probability of augmenting step.}

\item{pShort}{float, probability of shortening step.}

\item{seqsInit}{list of data frame, sequences from which start/continue MCMC
chain. They must have all the same length.}
}
\value{
permut, list of sequences.
}
\description{
Computes an MCMC chain and returns nmax sequences from the MCMC
chain.
}
