% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_parameters.R
\name{rubinsRule}
\alias{rubinsRule}
\title{Rubin´s rule for weighted data}
\usage{
rubinsRule(x, se, w)
}
\arguments{
\item{x}{values of the parameters}

\item{se}{standard errors of the parameters}

\item{w}{values of the weights}
}
\value{
value of the averaged mean and the standard error
}
\description{
Rubin´s rule for weighted data
}
