% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampleVec.R
\name{sampleVec}
\alias{sampleVec}
\title{Step for permuting two elements in a given sequence.}
\usage{
sampleVec(x, ...)
}
\arguments{
\item{x}{vector}
}
\value{
sample of elements of the vector
}
\description{
Step for permuting two elements in a given sequence.
}
