% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ABMCEM.R
\name{sigmaHat}
\alias{sigmaHat}
\title{Auxiliar function to compute \enq{\hat\sigma}.}
\usage{
sigmaHat(mt, loglikPrev, loglikCur, w)
}
\arguments{
\item{mt}{number of observations}

\item{loglikPrev}{log-likelihood at time \enq{t-1}}

\item{loglikCur}{log-likelihood at time \enq{t}}

\item{w}{weights (for importance sampling)}
}
\value{
estimator of \enq{\sigma^2}
}
\description{
Auxiliar function to compute \enq{\hat\sigma}.
}
\references{
Caffo, B. S., Jank, W., & Jones, G. L. (2005). Ascent-based Monte Carlo expectation–maximization. \emph{Journal of the Royal Statistical Society Series B: Statistical Methodology, 67(2)}, 235-251.
}
