% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simulation.R
\name{simulationDyNAM}
\alias{simulationDyNAM}
\title{Function that generates simulations of the process of competition between
creation and deletion of ties.
Indegree, outdegree, reciprocity and transitivity in the choice model.
Constant rate.}
\usage{
simulationDyNAM(
  nSimCrea,
  nSimDel,
  parmsChoiceCrea,
  parmsChoiceDel,
  endTime = 1,
  rep = FALSE,
  init_matrix = s501
)
}
\arguments{
\item{nSimCrea}{number of Creation events to be created (in average)}

\item{nSimDel}{number of Deletion events to be created (in average)}

\item{parmsChoiceCrea}{values of the Creation choice parameters (indeg,
outdeg, recip, trans)}

\item{parmsChoiceDel}{values of the Deletion choice parameters (indeg,
outdeg, recip, trans)}

\item{endTime}{time between networks, set to 1 by definition}

\item{rep}{allow for repetition of events with same sender-receiver pair}

\item{init_matrix}{adjacency matrix of initial network}
}
\value{
List with results of estimation given by goldfish, initial and final
networks (adjacency matrix), sequence of events (eventsMod)
}
\description{
Function that generates simulations of the process of competition between
creation and deletion of ties.
Indegree, outdegree, reciprocity and transitivity in the choice model.
Constant rate.
}
