% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{stepAugment}
\alias{stepAugment}
\title{Step for augmenting a given sequence by adding a tie \eqn{e} two times.}
\usage{
stepAugment(seq, tieNames, gammaEplus, gammaPlus, m, me, net0, pAug)
}
\arguments{
\item{seq}{data frame, sequence of events.}

\item{tieNames}{vector, labels of ties (e.g., "12" if the tie is from
sender 1 and receiver 2).}

\item{gammaEplus}{matrix, probabilities needed for the computation.}

\item{gammaPlus}{float, sum of gammaEplus.}

\item{m}{float, sum of me.}

\item{me}{matrix, probabilities needed for the computation.}

\item{net0}{matrix, initial network.}
}
\value{
list of
\item{sender}{sender of the new event.}
\item{receiver}{receiver of the new event.}
\item{place}{positions were the event was included.}
\item{typeA}{type of augmentation (two events together or separetly).}
\item{pDoStep}{probability of doing the step.}
}
\description{
Step for augmenting a given sequence by adding a tie \eqn{e} two times.
}
