% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{stepPerm}
\alias{stepPerm}
\title{Step for permuting two elements in a given sequence.}
\usage{
stepPerm(seq, tieNames, m, me)
}
\arguments{
\item{seq}{data frame, sequence of events.}

\item{tieNames}{vector, labels of ties (e.g., "12" if the tie is from
sender 1 and receiver 2).}

\item{m}{float, sum of me.}

\item{me}{matrix, probabilities needed for the computation.}
}
\value{
list of
\item{sender}{sender of the new event.}
\item{receiver}{receiver of the new event.}
\item{place}{positions were the event was included.}
\item{pDoStep}{probability of doing the step.}
}
\description{
Step for permuting two elements in a given sequence.
}
