% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{stepShort}
\alias{stepShort}
\title{Step for shortening a given sequence by removing a tie \eqn{e} in
two positions.}
\usage{
stepShort(seq, tieNames, gammaEminus, gammaMinus, m, me, Kel_g1, auxDf, pShort)
}
\arguments{
\item{seq}{data frame, sequence of events.}

\item{tieNames}{vector, labels of ties (e.g., "12" if the tie is from
sender 1 and receiver 2).}

\item{gammaEminus}{matrix, probabilities needed for the computation.}

\item{gammaMinus}{float, sum of gammaEplus.}

\item{m}{float, sum of me.}

\item{me}{matrix, probabilities needed for the computation.}

\item{auxDf}{data frame, from getKelMeMatrix}
}
\value{
list of
\item{sender}{sender of the new event.}
\item{receiver}{receiver of the new event.}
\item{place}{positions were the event was included.}
\item{typeS}{type of shortening (two events together or separetly).}
\item{pDoStep}{probability of doing the step.}
}
\description{
Step for shortening a given sequence by removing a tie \eqn{e} in
two positions.
}
