% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_timeGenerator.R
\name{timeGenerator}
\alias{timeGenerator}
\title{Time generator function.}
\usage{
timeGenerator(seq, nAct, theta)
}
\arguments{
\item{seq}{sequence of events}

\item{nAct}{number of actors}

\item{theta}{list with Crea and Del objects with rate parameters}
}
\value{
vector of estimated time values for each event in the sequence
}
\description{
Generate possible times for a sequence of events following the DyNAM model.
}
\examples{
library(RSiena)
X0 <- s501
X1 <- s502
seq <- EMPreprocessing(X0, X1)
timeGenerator(seq, ncol(X0), list(c("Crea" = rep(1, ncol(X0)), "Del" = rep(1, ncol(X0)))))
}
