% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ABMCEM.R
\name{MCEMalgorithmRateMCMC}
\alias{MCEMalgorithmRateMCMC}
\title{Ascent-Based Markov-Chain Expectation-Maximization algorithm with MCMC and rate}
\usage{
MCEMalgorithmRateMCMC(
  nmax,
  net0,
  net1,
  theta0,
  beta0,
  fixedparameters = c(NA, NA, NA, 20, 20),
  formula,
  initTime = 0,
  endTime = 1,
  num_cores = 1,
  errType2 = 0.8,
  alpha = 0.9,
  gamma = 0.9,
  thr = 0.001,
  maxIter = 1000,
  thin = 10,
  pShort = 0.35,
  pAug = 0.35,
  pPerm = 0.3,
  k_perm = 5,
  maxIterPT = 1000
)
}
\arguments{
\item{net0}{initial network}

\item{net1}{final network}

\item{theta0}{initial values for rate parameters}

\item{beta0}{initial values for effect parameters}

\item{formula}{formula of the choice model}

\item{initTime}{initial time. If formulaRate is not \code{NULL}, this parameter cannot be null.}

\item{endTime}{end time. If formulaRate is not \code{NULL}, this parameter cannot be null.}

\item{num_cores}{number of cores for paralelization}

\item{errType2}{value of power wanted. Default 0.8}

\item{alpha}{value for lower bound computation. Default 0.9}

\item{gamma}{value for stopping rule. Default 0.9}

\item{thr}{threshold for algorithm stop}

\item{maxIter}{maximum number of iterations for MCMC}

\item{thin}{number of MCMC steps beween samples}

\item{pShort}{probability of shortening the sequence (MCMC)}

\item{pAug}{probability of augmenting the sequence (MCMC)}

\item{pPerm}{probability of permuting two elements of the sequence (MCMC)}

\item{nmax0}{initial value of number of permuations for step 0}

\item{formulaRate}{formula of the rate model. Default value is \code{NULL}. Leave default for constant rate models (faster algorithm)}
}
\value{
list of
\item{logLik}{log likelihood of the sequences at the last step}
\item{beta}{estimation of effect parameters}
\item{se}{estimation of standard errors}
\item{index}{number of iterations until convergence of EM algorithm}
\item{diff}{value of stopping rule at last iteration}
\item{permut}{last set of permutations}
\item{betaCreaDF}{last data frame of goldfish estimators}
\item{betaDelDF}{last data frame of goldfish estimators}
}
\description{
This algorithm allows for an implementation of the ABMCEM algorithm with and without constant rates.
For the MCMC steps, three types of steps are done: shortening, augmenting and permutation of the sequence.
\eqn{pShort + pAug + pPerm = 1}
}
\references{
Caffo, B. S., Jank, W., & Jones, G. L. (2005). Ascent-based Monte Carlo expectation–maximization.
\emph{Journal of the Royal Statistical Society Series B: Statistical Methodology, 67(2)}, 235-251.

Stadtfeld, C., and Block, P. (2017). Interactions, Actors, and Time:
Dynamic Network Actor Models for Relational Events.
\emph{Sociological Science 4 (1)}, 318-52. \doi{10.15195/v4.a14}

Koskinen J.(2004). BAYESIAN INFERENCE FOR LONGITUDINAL SOCIAL NETWORKS.
\emph{Research Report, number 2004: 4.}.

Ruth, W. (2024). A review of Monte Carlo-based versions of the EM algorithm.
\emph{arXiv preprint} arXiv:2401.00945.

Snijders, T. A., Koskinen, J., & Schweinberger, M. (2010).
Maximum likelihood estimation for social network dynamics.
\emph{The annals of applied statistics, 4(2)}, 567.
}
