% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MCMC.R
\name{PT_Rate_MCMC}
\alias{PT_Rate_MCMC}
\title{PT_Rate_MCMC (simple version with augmentation/shortening/permutation)}
\usage{
PT_Rate_MCMC(
  nmax,
  nPT,
  seqsPT,
  H,
  actDfnodes,
  formula,
  net0,
  beta,
  theta,
  fixedparameters,
  initTime,
  endTime,
  burnIn = TRUE,
  burnInIter = 500,
  maxIter = 10000,
  thin = 50,
  T0 = 100,
  nStepExch = 10,
  pAug = 0.35,
  pShort = 0.35,
  pPerm = 0.3,
  k = 5,
  num_cores = num_cores,
  typeTemp = "sequential",
  r = 1/2,
  index = 1
)
}
\arguments{
\item{nmax}{integer, number of sequences to be returned.}

\item{H}{integer, hamming distance between net0 and net1.}

\item{actDfnodes}{object of type ´nodes´ from goldfish.}

\item{formula}{formula, the one used in goldfish model.}

\item{net0}{matrix, initial observed network.}

\item{beta}{list, estimator of parameters of the model (creation and deletion).}

\item{burnIn}{boolean, indicates if burn in must be performed.}

\item{maxIter}{integer, maximum number of steps of the MCMC chain.}

\item{thin}{integer, number of steps between selected sequences.}

\item{pAug}{float, probability of augmenting step.}

\item{pShort}{float, probability of shortening step.}

\item{pPerm}{float, probability of permutation step.}

\item{seq}{data frame, sequence from which start/continue MCMC chain.}
}
\value{
permut, list of sequences.
}
\description{
Computes an MCMC chain and returns nmax sequences from the MCMC chain.
}
